Split apart the grammar tests
