+++ test/core/parser/grammar/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the sqlfluff.core.parser.grammar module."""
